requirements.txt
python-dotenv
mysql-connector-python
pymysql
google-generativeai
streamlit
langchain
langchain-community
langchain-core
langchain-experimental
